requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backpack==0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.6.8
cloudinary==1.29.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
facebook-sdk==3.1.0
Faker==0.8.18
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
inflection==0.3.1
lazy-object-proxy==1.7.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.19.4
psycopg2==2.9.3
pyaml==16.12.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pylev==1.4.0
python-dateutil==2.8.2
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.2
toml==0.10.2
twilio==7.8.2
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==1.5.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.0.0
wrapt==1.13.3
